+++ projects/ngx-scrollbar-demo/src/app/integration-page/integration-page.component.ts
@@ -2,7 +2,7 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { ExampleNgxDatatableComponent } from '../example-ngx-datatable/example-ngx-datatable.component';
 import { ExampleInfiniteScrollComponent } from '../example-infinite-scroll/example-infinite-scroll.component';
 import { ExampleAgGridTableComponent } from '../example-ag-grid-table/example-ag-grid-table.component';
-import { PrimeNgTableComponent } from '../prime-ng-table/prime-ng-table.component';
+import { PrimeNgComponent } from '../prime-ng-table/prime-ng.component';
 
 @Component({
   selector: 'app-integration-page',
@@ -11,7 +11,7 @@ import { PrimeNgTableComponent } from '../prime-ng-table/prime-ng-table.componen
     ExampleNgxDatatableComponent,
     ExampleInfiniteScrollComponent,
     ExampleAgGridTableComponent,
-    PrimeNgTableComponent
+    PrimeNgComponent
   ],
   templateUrl: './integration-page.component.html',
   styleUrl: './integration-page.component.scss',